% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_backend.R
\name{big_backend}
\alias{big_backend}
\title{Set or get the preferred backend}
\usage{
big_backend(backend = NULL)
}
\arguments{
\item{backend}{Set the backend ("data.table", "arrow", "duckdb")}
}
\value{
The current backend
}
\description{
Set or get the preferred backend
}
