% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_chunk_map.R
\name{big_chunk_map}
\alias{big_chunk_map}
\title{Map a function over dataset chunks and write results to a file}
\usage{
big_chunk_map(file, FUN, chunk_size = 10000, output_file)
}
\arguments{
\item{file}{Path to the input CSV file}

\item{FUN}{Function to apply to each chunk}

\item{chunk_size}{Number of rows per chunk}

\item{output_file}{Path to save the transformed file}
}
\description{
Map a function over dataset chunks and write results to a file
}
