% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_filter.R
\name{big_filter}
\alias{big_filter}
\title{Filter rows from a large dataset efficiently}
\usage{
big_filter(data, ...)
}
\arguments{
\item{data}{A dataframe or a database connection}

\item{...}{Filter conditions}
}
\value{
The filtered dataset
}
\description{
Filter rows from a large dataset efficiently
}
