% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_group_by.R
\name{big_group_by}
\alias{big_group_by}
\title{Group a large dataset efficiently}
\usage{
big_group_by(data, ...)
}
\arguments{
\item{data}{A dataframe or DuckDB connection}

\item{...}{Columns to group by}
}
\value{
A grouped dataframe or DuckDB query result
}
\description{
Group a large dataset efficiently
}
