% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_optimize.R
\name{big_optimize}
\alias{big_optimize}
\title{Suggest optimizations based on dataset size}
\usage{
big_optimize(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
Suggested optimization
}
\description{
Suggest optimizations based on dataset size
}
