% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_parallel_mutate.R
\name{big_parallel_mutate}
\alias{big_parallel_mutate}
\title{Parallelized version of mutate()}
\usage{
big_parallel_mutate(data, ..., n_cores = parallel::detectCores())
}
\arguments{
\item{data}{A dataframe}

\item{...}{Mutate expressions}

\item{n_cores}{Number of CPU cores to use for parallel execution.}
}
\value{
Mutated dataframe
}
\description{
Parallelized version of mutate()
}
