% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_parallel_summarize.R
\name{big_parallel_summarize}
\alias{big_parallel_summarize}
\title{Parallelized version of summarize()}
\usage{
big_parallel_summarize(data, ..., n_cores = parallel::detectCores())
}
\arguments{
\item{data}{A dataframe}

\item{...}{Summarization expressions}

\item{n_cores}{Number of CPU cores to use for parallel execution.}
}
\value{
Summarized dataframe
}
\description{
Parallelized version of summarize()
}
\examples{
\dontrun{
# Basic parallel summarization
result <- big_parallel_summarize(data,
  mean_val = mean(value),
  sd_val = sd(value),
  count = n()
)

# Complex calculations in parallel
result <- big_parallel_summarize(data,
  quantiles = quantile(value, probs = seq(0, 1, 0.1)),
  correlation = cor(value1, value2),
  custom_stat = expensive_stat_function(value)
)

# Specify number of cores
result <- big_parallel_summarize(data,
  bootstrap_mean = mean(replicate(1000, mean(sample(value, replace = TRUE)))),
  n_cores = 4
)

# Combining with other operations
result <- data \%>\%
  big_filter(value > 0) \%>\%
  big_group_by(category) \%>\%
  big_parallel_summarize(
    mean_val = mean(value),
    sd_val = sd(value),
    q95 = quantile(value, 0.95)
  )
}
}
