% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_read_parquet.R
\name{big_read_parquet}
\alias{big_read_parquet}
\title{Read a large Parquet file using an optimized backend}
\usage{
big_read_parquet(file, backend = "auto")
}
\arguments{
\item{file}{Path to the Parquet file}

\item{backend}{Choose backend: "auto", "arrow", "duckdb"}
}
\value{
A dataframe
}
\description{
Read a large Parquet file using an optimized backend
}
