% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_summarize.R
\name{big_summarize}
\alias{big_summarize}
\title{Summarize a large dataset efficiently}
\usage{
big_summarize(data, ...)
}
\arguments{
\item{data}{A dataframe or DuckDB connection}

\item{...}{Summarization expressions}
}
\value{
A summarized dataframe
}
\description{
Summarize a large dataset efficiently
}
