% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_write_csv.R
\name{big_write_csv}
\alias{big_write_csv}
\title{Write a large dataset to a CSV efficiently}
\usage{
big_write_csv(data, file, backend = "auto", chunk_size = NULL)
}
\arguments{
\item{data}{A dataframe to write}

\item{file}{Path to save the file}

\item{backend}{Choose backend: "auto", "data.table", "arrow", "duckdb"}

\item{chunk_size}{Optional chunk size for writing in parts}
}
\description{
Write a large dataset to a CSV efficiently
}
