% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_write_parquet.R
\name{big_write_parquet}
\alias{big_write_parquet}
\title{Write a large dataset to a Parquet file efficiently}
\usage{
big_write_parquet(data, file, backend = "auto")
}
\arguments{
\item{data}{A dataframe to write}

\item{file}{Path to save the Parquet file}

\item{backend}{Choose backend: "auto", "arrow", "duckdb"}
}
\description{
Write a large dataset to a Parquet file efficiently
}
