% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigr-package.R
\docType{package}
\name{bigr-package}
\alias{bigr}
\alias{bigr-package}
\title{R Markdown Template for 'big'}
\description{
'big' (\url{https://tmcw.github.io/big/demo.html}) is a presentation for
 busy messy hackers.
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com}

}
