% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{bigr}
\alias{bigr}
\alias{bigr-package}
\title{small bigr package}
\description{
This package is designed for easy life of BIG company.
}
\details{
It loads some popular packages by default.
To see the list of available functions click on index below!
Be happy, don't worry :)
}
\author{
Boris Demeshev
}

