% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADC_conversion.R
\name{compare}
\alias{compare}
\title{Get SNP positions, reference and alternative alleles based on the reference
Align alternatives to reference and discard low score alignment tags
Discard tags if alternative in the target locus is N
Do the complement reverse if cloneID present in the botloci vector}
\usage{
compare(one_tag, botloci)
}
\arguments{
\item{one_tag}{madc file split by tag}

\item{botloci}{file containing the target IDs that were designed in the bottom strand}
}
\description{
Get SNP positions, reference and alternative alleles based on the reference
Align alternatives to reference and discard low score alignment tags
Discard tags if alternative in the target locus is N
Do the complement reverse if cloneID present in the botloci vector
}
