% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADC_conversion.R
\name{create_VCF_body}
\alias{create_VCF_body}
\title{Creates VCF body from CSV generated by loop_though_dartag_report}
\usage{
create_VCF_body(
  csv,
  rm_multiallelic_SNP = TRUE,
  multiallelic_SNP_dp_thr = 10,
  n.cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{csv}{CSV file generated by loop_though_dartag_report}

\item{rm_multiallelic_SNP}{logical. If TRUE, SNP with more than one alternative base will be removed. If FALSE, check \code{multiallelic_SNP_dp_thr} specs}

\item{multiallelic_SNP_dp_thr}{numerical. If \code{rm_multiallelic_SNP} is FALSE, set a total depth by tag threshold to eliminate low frequency SNP allele. If the threshold does not eliminate the multiallelic aspect of the marker, the marker is discarded. This is likely to happen to paralogous sites.}

\item{n.cores}{number of cores to be used in the parallelization}

\item{verbose}{print metrics on the console}
}
\description{
Creates VCF body from CSV generated by loop_though_dartag_report
}
