% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ct_luckyguy}
\alias{ct_luckyguy}
\title{Lucky guy function}
\usage{
ct_luckyguy(user_ans, etalon, ...)
}
\arguments{
\item{user_ans}{the vector of user responces}

\item{etalon}{the vector of etalon cathegories}

\item{translit}{logical whether to do transliteration}
}
\value{
pretty correspondance table if you are a lucky guy
}
\description{
This function tries to create a final correspondance table
starting from a vector of user answers and a vector of etalons
}
\details{
The Levenstein distance is used. When to matches give the same distance
the first one is returned. Warning: for lucky guys only!
}
\examples{
ct_luckyguy(user_ans,etalon,translit = FALSE)
}

