% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixMADC.R
\name{fixMADC}
\alias{fixMADC}
\title{Fix MADC File Allele IDs}
\usage{
fixMADC(madc.file, marker.file, n.summary.columns = NULL, output.file = NULL)
}
\arguments{
\item{madc.file}{Path to the MADC file to be filtered}

\item{marker.file}{Path to the three column marker ID file.
\itemize{
\item The first column is the existing list of unique CloneIDs (obtained from raw MADC CloneID),
where each row is a unique CloneID.
\item The second column is the chromosome that the marker is located (ie Chr01). No special characters (*#_-!.) are permitted in the
chromosome name.
\item The third column is the numeric position of the marker within the chromosome (ie 1234). No special characters (*#_-!.) are permitted.
}}

\item{n.summary.columns}{(optional) Number of summary columns to remove from MADC file not including the first three. Otherwise, the columns will be automatically detected and removed.}

\item{output.file}{Path to save the fixed allele ID MADC file (if NULL, data will not be saved)}
}
\value{
data.frame or saved csv file
}
\description{
Process raw MADC files to format and update the allele IDs with user supplied Chr and Pos information
}
\details{
This function can process raw MADC files to update the Allele IDs and Clone IDs to the Chr_Pos format with a user supplied file.
The output MADC will be the standard fixed allele ID format to support use in madc2vcf and BIGapp functions.
}
\examples{
#Example

#Example MADC
madc_file <- system.file("iris_DArT_MADC.csv", package="BIGr")
marker_file <- system.file("iris_MADC_marker_file.csv", package="BIGr")

#Fix the raw MADC file IDs to use the user provided Chr_Pos format
fixedMADC_df <- fixMADC(madc.file = madc_file,
                         marker.file = marker_file,
                         n.summary.columns = NULL,
                         output.file = NULL)



}
