% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADC_conversion.R
\name{get_OffTargets}
\alias{get_OffTargets}
\title{Converts MADC file to VCF recovering target and off-target SNPs}
\usage{
get_OffTargets(
  madc = NULL,
  botloci = NULL,
  hap_seq = NULL,
  n.cores = 5,
  rm_multiallelic_SNP = FALSE,
  multiallelic_SNP_dp_thr = 0,
  multiallelic_SNP_sample_thr = 0,
  out_vcf = NULL,
  verbose = TRUE
)
}
\arguments{
\item{madc}{path to MADC file}

\item{botloci}{path to file containing the target IDs that were designed in the bottom strand}

\item{hap_seq}{path to haplotype DB fasta file}

\item{n.cores}{number of cores to be used in the parallelization}

\item{rm_multiallelic_SNP}{logical. If TRUE, SNP with more than one alternative base will be removed. If FALSE, check \code{multiallelic_SNP_dp_thr} specs}

\item{multiallelic_SNP_dp_thr}{nnumerical. If \code{rm_multiallelic_SNP} is FALSE, set a minimum depth by tag threshold \code{multiallelic_SNP_dp_thr} combined with minimum number of samples \code{multiallelic_SNP_sample_thr} to eliminate low frequency SNP allele. If the threshold does not eliminate the multiallelic aspect of the marker, the marker is discarded. This is likely to happen to paralogous sites.}

\item{multiallelic_SNP_sample_thr}{numerical. If \code{rm_multiallelic_SNP} is FALSE, set a minimum depth by tag threshold combined with minimum number of samples \code{multiallelic_SNP_sample_thr} to eliminate low frequency SNP allele. If the threshold does not eliminate the multiallelic aspect of the marker, the marker is discarded. This is likely to happen to paralogous sites.}

\item{out_vcf}{output VCF file name}

\item{verbose}{print metrics on the console}
}
\description{
Converts MADC file to VCF recovering target and off-target SNPs
}
