% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OffTargets.R
\name{get_ref_alt_hap_seq}
\alias{get_ref_alt_hap_seq}
\title{Converts the fasta to a data.frame with first column the AlleleID and and second the AlleleSequence
The function will work even if the sequence is split in multiple lines}
\usage{
get_ref_alt_hap_seq(hap_seq)
}
\arguments{
\item{hap_seq}{haplotype db}
}
\description{
Converts the fasta to a data.frame with first column the AlleleID and and second the AlleleSequence
The function will work even if the sequence is split in multiple lines
}
