% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADC_conversion.R
\name{loop_though_dartag_report}
\alias{loop_though_dartag_report}
\title{Include SNP_position_in_Genome, Ref, and Alt information}
\usage{
loop_though_dartag_report(
  report,
  botloci,
  hap_seq,
  n.cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{report}{MADC file}

\item{botloci}{file containing the target IDs that were designed in the bottom strand}

\item{hap_seq}{haplotype DB fasta file}

\item{n.cores}{number of cores to be used in the parallelization}

\item{verbose}{print metrics on the console}
}
\description{
Include SNP_position_in_Genome, Ref, and Alt information
}
