% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADC_conversion.R
\name{merge_counts}
\alias{merge_counts}
\title{Function made for parallelization of create_VCF_body function}
\usage{
merge_counts(
  cloneID_unit,
  rm_multiallelic_SNP = FALSE,
  multiallelic_SNP_dp_thr = 0
)
}
\arguments{
\item{cloneID_unit}{one item of csv file split by cloneID}

\item{rm_multiallelic_SNP}{logical. If TRUE, SNP with more than one alternative base will be removed. If FALSE, check \code{multiallelic_SNP_dp_thr} specs}

\item{multiallelic_SNP_dp_thr}{numerical. If \code{rm_multiallelic_SNP} is FALSE, set a total depth by tag threshold to eliminate low frequency SNP allele. If the threshold does not eliminate the multiallelic aspect of the marker, the marker is discarded. This is likely to happen to paralogous sites.}
}
\description{
Function made for parallelization of create_VCF_body function
}
