% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_MADCs.R
\name{merge_MADCs}
\alias{merge_MADCs}
\title{Merge MADC files}
\usage{
merge_MADCs(..., out_madc = NULL)
}
\arguments{
\item{...}{one or more MADC files path}
}
\description{
If duplicated samples exist in different files, a extension (e.g."_file:1") will
be added at the end of the sample name
}
