% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breedtools_functions.R
\name{QPsolve}
\alias{QPsolve}
\title{Performs whole genome breed composition prediction.}
\usage{
QPsolve(Y, X)
}
\arguments{
\item{Y}{numeric vector of genotypes (with names as SNPs) from a single animal.
coded as dosage of allele B \code{{0, 1, 2, ..., ploidy}}}

\item{X}{numeric matrix of allele frequencies from reference animals}

\item{p}{numeric indicating number of breeds represented in X}

\item{names}{character names of breeds}
}
\value{
data.frame of breed composition estimates
}
\description{
Performs whole genome breed composition prediction.
}
\references{
Funkhouser SA, Bates RO, Ernst CW, Newcom D, Steibel JP. Estimation of genome-wide and locus-specific
breed composition in pigs. Transl Anim Sci. 2017 Feb 1;1(1):36-44.
}
