% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_csv}
\alias{read_csv}
\title{Modified read.csv with stringsAsFactors = FALSE}
\usage{
read_csv(...)
}
\arguments{
\item{...}{parameters passed to read.csv}
}
\value{
data.frame
}
\description{
Modified read.csv with stringsAsFactors = FALSE
}
\examples{
read_csv("myfile.csv")
}

