% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_orderwords}
\alias{str_orderwords}
\title{Order words in a vector of sentences}
\usage{
str_orderwords(x, sep = " ")
}
\arguments{
\item{x}{the vector of sentences}
}
\value{
a vector of sentences with ordered words
}
\description{
This function order words in each sentence of a character vector
}
\examples{
str_orderwords(c("привет","Маша, это я, Дубровский"))
}

