% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_sent2words}
\alias{str_sent2words}
\title{Convert vector of sentences to a data.frame of separate words}
\usage{
str_sent2words(x, cleanup = TRUE)
}
\arguments{
\item{x}{the vector of sentences}

\item{cleanup}{logical indicates whether to remove punctuation and some other cleanup}
}
\value{
data.frame with 3 variables: word, word_n (number of word), sent_n (number of sentence)
}
\description{
Convert vector of sentences to a data.frame of separate words
The output is a data.frame with 3 variables:
word, word_n (number of word), sent_n (number of sentence)
}
\examples{
str_sent2words(c("привет","Маша, это я, Дубровский"))
}

