% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_stand}
\alias{str_stand}
\title{Standartize character vector}
\usage{
str_stand(z)
}
\arguments{
\item{x}{the character vector}
}
\value{
standartized character vector
}
\description{
Standartize means: trimming white space at the end and in the beginning,
elimination of punctiation, transliteration, removing capitalization
}
\examples{
str_stand("пРивет!")
}

