% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_translit}
\alias{str_translit}
\title{Transliterate cyrillic text}
\usage{
str_translit(x)
}
\arguments{
\item{x}{the vector of cyrillic characters}
}
\value{
transliterated vector
}
\description{
This function uses the transliteration tradition where "й" goes to "y"
}
\examples{
str_translit("привет")
}

