% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_word}
\alias{str_word}
\title{Get specific word from a vector of character sentences}
\usage{
str_word(x, num = 1)
}
\arguments{
\item{x}{the vector of cyrillic characters}

\item{num}{the number of word, negative numbers mean "from the end of sentence"}
}
\value{
data frame of words with numbers
}
\description{
This function uses the transliteration tradition where "й" goes to "y"
}
\examples{
str_word(c("привет","Маша, это я, Дубровский"))
}

