% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog2vcf.R
\name{updog2vcf}
\alias{updog2vcf}
\title{Export Updog Results as VCF}
\usage{
updog2vcf(multidog.object, ploidy, output.file)
}
\arguments{
\item{multidog.object}{updog output object with class "multidog" from dosage calling}

\item{ploidy}{The ploidy of the species being analyzed}

\item{output.file}{output file name and path}
}
\value{
A vcf file
}
\description{
This function will convert an Updog output to a VCF file
}
\details{
When performing dosage calling for multiple SNPs using Updog, the output file contains information for all loci and all samples.
This function will convert the updog output file to a VCF file, while retaining the information for the values that are commonly
used to filter low quality and low confident dosage calls.
}
\references{
Updog R package
}
