% Generated by roxygen2 (4.0.2): do not edit by hand
\name{utf2cp}
\alias{utf2cp}
\title{Change encoding of cyrillic text from "utf8" to "cp1251"}
\usage{
utf2cp(x)
}
\arguments{
\item{x}{the vector or data.frame of cyrillic characters}
}
\value{
reencoded vector or data.frame
}
\description{
This function changes encoding of a vector or data.frame from "utf8" to "cp1251"
}
\examples{
utf2cp("привет")
}

