% Generated by roxygen2 (4.0.2): do not edit by hand
\name{utf2cp0}
\alias{utf2cp0}
\title{Change encoding of cyrillic text from "utf8" to "cp1251"}
\usage{
utf2cp0(x)
}
\arguments{
\item{x}{the vector of cyrillic characters}
}
\value{
reencoded vector
}
\description{
This function changes encoding of a vector from "utf8" to "cp1251"
This function applies changes only to character and factor variables
}
\examples{
utf2cp0("привет")
}

