% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breedtools_functions.R
\name{allele_freq_poly}
\alias{allele_freq_poly}
\title{Computes allele frequencies for specified populations given SNP array data}
\usage{
allele_freq_poly(geno, populations, ploidy = 2)
}
\arguments{
\item{geno}{matrix of genotypes coded as the dosage of allele B \code{{0, 1, 2, ..., ploidy}}
with individuals in rows (named) and SNPs in columns (named)}

\item{populations}{list of named populations. Each population has a vector of IDs
that belong to the population. Allele frequencies will be derived from all animals}

\item{ploidy}{integer indicating the ploidy level (default is 2 for diploid)}
}
\value{
data.frame consisting of allele_frequencies for populations (columns) for
each SNP (rows)
}
\description{
Computes allele frequencies for specified populations given SNP array data
}
\references{
Funkhouser SA, Bates RO, Ernst CW, Newcom D, Steibel JP. Estimation of genome-wide and locus-specific
breed composition in pigs. Transl Anim Sci. 2017 Feb 1;1(1):36-44.
}
