% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_arrange.R
\name{big_arrange}
\alias{big_arrange}
\title{Sort a large dataset efficiently}
\usage{
big_arrange(data, ...)
}
\arguments{
\item{data}{A dataframe or DuckDB connection}

\item{...}{Columns to sort by}
}
\value{
A sorted dataframe
}
\description{
Sort a large dataset efficiently
}
