% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_memory_usage.R
\name{big_memory_usage}
\alias{big_memory_usage}
\title{Calculate memory usage of a dataset}
\usage{
big_memory_usage(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
Memory usage in MB
}
\description{
Calculate memory usage of a dataset
}
