% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_mutate.R
\name{big_mutate}
\alias{big_mutate}
\title{Mutate columns in a large dataset efficiently}
\usage{
big_mutate(data, ...)
}
\arguments{
\item{data}{A dataframe or a database connection}

\item{...}{Mutate expressions}
}
\value{
The mutated dataset
}
\description{
Mutate columns in a large dataset efficiently
}
