% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_parallel_apply.R
\name{big_parallel_apply}
\alias{big_parallel_apply}
\title{Apply a function to a dataset using parallel computing}
\usage{
big_parallel_apply(data, FUN, n_cores = parallel::detectCores())
}
\arguments{
\item{data}{A dataframe}

\item{FUN}{Function to apply}

\item{n_cores}{Number of CPU cores to use}
}
\value{
Processed dataframe
}
\description{
Apply a function to a dataset using parallel computing
}
