% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_parallel_mutate.R
\name{big_parallel_mutate}
\alias{big_parallel_mutate}
\title{Parallelized version of mutate()}
\usage{
big_parallel_mutate(data, ..., n_cores = parallel::detectCores())
}
\arguments{
\item{data}{A dataframe}

\item{...}{Mutate expressions}

\item{n_cores}{Number of CPU cores to use for parallel execution.}
}
\value{
Mutated dataframe
}
\description{
Parallelized version of mutate()
}
\examples{
\dontrun{
# Basic parallel mutation
result <- big_parallel_mutate(data,
  new_col = expensive_calculation(value)
)

# Multiple parallel mutations
result <- big_parallel_mutate(data,
  squared = value^2,
  logged = log(value),
  normalized = scale(value)
)

# Specify number of cores
result <- big_parallel_mutate(data,
  complex_calc = expensive_function(col1, col2),
  n_cores = 4
)

# Combining with other operations
result <- data \%>\%
  big_filter(value > 0) \%>\%
  big_parallel_mutate(
    log_value = log(value),
    z_score = scale(value)
  )
}
}
