% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_parallel_summarize.R
\name{big_parallel_summarize}
\alias{big_parallel_summarize}
\title{Parallelized version of summarize()}
\usage{
big_parallel_summarize(data, ..., n_cores = parallel::detectCores())
}
\arguments{
\item{data}{A dataframe}

\item{...}{Summarization expressions}
}
\value{
Summarized dataframe
}
\description{
Parallelized version of summarize()
}
