% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_select.R
\name{big_select}
\alias{big_select}
\title{Select columns from a large dataset efficiently}
\usage{
big_select(data, ...)
}
\arguments{
\item{data}{A dataframe or a database connection}

\item{...}{Columns to select}
}
\value{
The selected columns
}
\description{
Select columns from a large dataset efficiently
}
