% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{bigr}
\alias{.onLoad}
\alias{bigr}
\alias{bigr-package}
\title{bigr}
\usage{
.onLoad(libname = find.package("bigr"), pkgname = "bigr")
}
\description{
This package is designed for BIG company. It loads some popular packages by default.
To see the list of available functions click on index below!
Be happy, don't worry :)
}
\author{
Boris Demeshev
}

