% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MAF.R
\name{calculate_MAF}
\alias{calculate_MAF}
\title{Calculate Minor Allele Frequency from a Genotype Matrix}
\usage{
calculate_MAF(df, ploidy)
}
\arguments{
\item{df}{Genotype matrix or data.frame}

\item{ploidy}{The ploidy of the species being analyzed}
}
\value{
A dataframe of AF and MAF values for each marker
}
\description{
This function calculates the allele frequency and minor allele frequency from a genotype matrix.
It assumes that the Samples are the columns, and the genomic markers are in rows. Missing data should
be set as NA, which will then be ignored for the calculations. All samples must have the same ploidy.
}
