% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ped.R
\name{check_ped}
\alias{check_ped}
\title{Evaluate Pedigree File for Accuracy}
\usage{
check_ped(ped.file)
}
\arguments{
\item{ped.file}{path to pedigree text file}
}
\value{
A cleaned pedigree .txt file
}
\description{
clean_pedigree takes a 3-column pedigree tab separated file with columns labeled as id sire
dam in any order and checks for ids that appear more than once in the id
column, ids that appear as both sire and dam, and circular dependencies where
an id is its own sire or dam.
}
\details{
This function also looks for any sires or dams that
are not included in the pedigree in the id column and adds them with unknown parents.
}
