% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMADC.R
\name{filterMADC}
\alias{filterMADC}
\title{Filter MADC Files}
\usage{
filterMADC(
  madc_file,
  min.mean.reads = NULL,
  max.mean.reads = NULL,
  max.match.mhaps = 10,
  min.reads.per.site = NULL,
  min.ind.with.reads = NULL,
  target_only = FALSE,
  fixed_allele_ids = FALSE,
  plot.summary = FALSE,
  output.file = NULL
)
}
\arguments{
\item{madc_file}{Path to the MADC file to be filtered}

\item{min.mean.reads}{Minimum mean read depth for filtering}

\item{max.mean.reads}{Maximum mean read depth for filtering}

\item{max.match.mhaps}{Maximum number of matching mhaps per target loci}

\item{min.reads.per.site}{Minimum number of reads per site for filtering}

\item{min.ind.with.reads}{Minimum number of individuals with reads for filtering}

\item{target_only}{Logical indicating whether to filter for target loci only}

\item{fixed_allele_ids}{Logical indicating whether the MADC file has been pre-processed for fixed allele IDs}

\item{plot.summary}{Logical indicating whether to plot summary statistics}

\item{output.file}{Path to save the filtered data (if NULL, data will not be saved)}

\item{verbose}{Logical indicating whether to print additional information during processing}
}
\value{
data.frame or saved csv file
}
\description{
Filter and process MADC files to remove low quality microhaplotypes
}
\details{
This function can filter raw MADC files or pre-processed MADC files with fixed allele IDs. Additionally,
it can filter based on mean read depth, number of mhaps per target loci, and other criteria. Optionally, users
can scale and normalize the data in preparation for conversion to relationship matrices,
plot summary statistics, and save the filtered data to a file.
}
\examples{
#Example...

##Plots
#Mean read depth
#Number of Altmatch and Refmatch mhaps per target loci


}
