% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_concordance.R
\name{imputation_concordance}
\alias{imputation_concordance}
\title{Calculate Concordance between Imputed and Reference Genotypes}
\usage{
imputation_concordance(
  reference_genos,
  imputed_genos,
  missing_code = NULL,
  snps_2_exclude = NULL,
  output_df_name = "imputation_concordance"
)
}
\arguments{
\item{reference_genos}{Genotype data.frame with rows as samples and columns as markers. Dosage recommended.}

\item{imputed_genos}{Genotype data.frame with rows as samples and columns as markers. Dosage recommended.}

\item{missing_code}{Optional input to consider missing data to exclude in concordance calculation.}

\item{snps_2_exclude}{Optional input to exclude specific markers from concordance calculation. Single column of marker ids.}

\item{output_df_name}{Optional input to assign the output dataframe to a specific variable name. Default is "imputation_concordance"}
}
\value{
2 outputs: 1) A data frame with sample IDs and concordance percentages. 2) A summary of concordance percentages.
}
\description{
This calculates the concordance between imputed and reference genotypes. It assumes that samples are rows and markers are columns.
It is recommended to use allele dosages (0,1,2) but will work with other formats. Missing data in reference or imputed genotypes
will not be considered for concordance if argument missing_code used. If a specific subset of markers should it can be provided as argument snps_2_exclude.
}
