% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madc2vcf.R
\name{madc2vcf}
\alias{madc2vcf}
\title{Format MADC Target Loci Read Counts Into VCF}
\usage{
madc2vcf(madc_file, output.file, get_REF_ALT = FALSE)
}
\arguments{
\item{madc_file}{Path to MADC file}

\item{output.file}{output file name and path}

\item{get_REF_ALT}{if TRUE recovers the reference and alternative bases by comparing the sequences. If more than one polymorphism are found for a tag, it is discarded.}
}
\value{
A VCF file v4.3 with the target marker read count information
}
\description{
This function will extract the read count information from a MADC file and convert to VCF file format.
}
\details{
The DArTag MADC file format is not commonly supported through existing tools. This function
will extract the read count information from a MADC file and convert it to a VCF file format for the
genotyping panel target markers only
}
\references{
Updog R package
}
